futu-openapi/requirements.txt
futu-api>=7.0.0
pandas>=1.3.0
